requirements.txt
einops>=0.6
tqdm>=4.64
Pillow>=9.2
numpy>=1.23
scikit-learn>=1.2
opencv-python>=4.7
matplotlib>=3.7